UpcomingScreenings - Separated sorting function to common util to share with other upcoming components
